api/requirements.txt
flask==3.0.0
sentence-transformers==2.3.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
